Move over some tests to their respective sub-package paths.
