+++ src/test/scala/nl/tudelft/tydi_chisel/examples/pipeline/PipelineExamplePlusTest.scala
@@ -1,12 +1,12 @@
-package nl.tudelft.tydi_chisel
+package nl.tudelft.tydi_chisel.examples.pipeline
 
 import chisel3._
 import chisel3.experimental.BundleLiterals.AddBundleLiteralConstructor
 import chisel3.experimental.VecLiterals.{AddObjectLiteralConstructor, AddVecLiteralConstructor}
 import chiseltest._
-import org.scalatest.flatspec.AnyFlatSpec
 import nl.tudelft.tydi_chisel.Conversions._
-import nl.tudelft.tydi_chisel.examples.pipeline._
+import nl.tudelft.tydi_chisel.{TydiProcessorTestWrapper, TydiTestWrapper}
+import org.scalatest.flatspec.AnyFlatSpec
 
 class PipelineExamplePlusTest extends AnyFlatSpec with ChiselScalatestTester {
   behavior of "PipelineExamplePlus"
