+++ src/test/scala/nl/tudelft/tydi_chisel/utils/LastSeqProcessorTest.scala
@@ -1,11 +1,10 @@
-package nl.tudelft.tydi_chisel.complexity_converter
+package nl.tudelft.tydi_chisel.utils
 
 import chisel3._
 import chisel3.experimental.VecLiterals.AddObjectLiteralConstructor
 import chiseltest._
-import org.scalatest.flatspec.AnyFlatSpec
 import nl.tudelft.tydi_chisel.printUtils._
-import nl.tudelft.tydi_chisel.utils.LastSeqProcessor
+import org.scalatest.flatspec.AnyFlatSpec
 
 class LastSeqProcessorTest extends AnyFlatSpec with ChiselScalatestTester {
   def UIntVec(elems: Int*): Vec[UInt] = {
